Changes for 4.6.0: configure a new Makefile for Win 95 formatting in many     rpc2 files ctest: had a bug: ping many retransmitted the opcode     alternatigly in host and net order venus-setup: /dev/cfs0 is not     readable anymore for others venus.version.h: wow! We are finally     updating this thing. venusvol.cc, vol_reintegrate: formatting vproc.h:     adaptation of CRTOUID for Linux to deal correctly with Samba exports.     BSD needs to make some decisions here. srvproc2.cc: NewConnectFS deals     well with restarting Venus now. clientproc.cc: minor C issue. type     changes in volutil Makerules: can install windows 95 client binaries     into a directory now New directories with djgpp tools activated under     djgpp. Linux 2.1 kernel: segfault issue finally solved! Better Ctrl-C     code; dealt with /proc/fs. Still problems left in /proc/fs/coda. Now     very systematic error returns with -signs removed.

+++ coda-src/vice/clientproc.cc
@@ -218,7 +218,7 @@ PRIVATE HostTable *client_GetVenusId(RPC2_Handle RPCid)
 
 	/* Look for a corresponding host entry. */
 	for (i = 0; i < maxHost; i++)
-		if (hostTable[i].host == (unsigned int) peer.RemoteHost.Value.InetAddress &&
+		if (hostTable[i].host == peer.RemoteHost.Value.InetAddress &&
 		    hostTable[i].port == peer.RemotePortal.Value.InetPortNumber)
 			break;
 	
@@ -230,7 +230,6 @@ PRIVATE HostTable *client_GetVenusId(RPC2_Handle RPCid)
 		hostTable[i].FirstClient = 0;
 		hostTable[i].id = 0;
 		sprintf(hostTable[i].HostName, "%08x", htonl(hostTable[i].host));
-
 		Lock_Init(&hostTable[i].lock);
 		maxHost++;
 	}
